rename: 이름 변경
- 오타 수정(constatns 폴더명을 constants로 변경)
